Successful login spawns form2, unsuccessful spawns error
A successful login will result in the main login form closing and the
next form opening. For now is just a blank form saying successful login.
If login is pressed and it is not successful, an error message will be
displayed at the top of the form. If either password or username is left
blank, a faint crimson glow will outline the blank edit box.