Month Read Finished
Month read should be finished besides perhaps some layout changes (i.e.
move a heading to another cell in the grid). Note the cells in the top
left are extra statistics that could be placed outside the grid if it
makes more sense. Now the rest of read should be straight forward by
reusing monthly read for weekly, arbitrary date range, and yearly. Then
just need to add the ability to change filters on the side next to the
grid. Then work out the other buttons at the bottom (need to be added).