Bug fixes and added two text files and edited todo
Added two text files called 1) need_todo_before_give_dad and 2)
need_to_add_to_db. The first lists all the major things we should try to
do before giving it to Max's dad. The second is a list I am creating
mostly for my own reference and includes things to adjust in the db in
order to make my code more modular in the future for the future. Also
fixed a few bugs and edited the todo list. Finally, changed the
calendars to start with Saturday instead of Sunday to mimic dlr